refactor: all size_ts

+++ include/vector.h
@@ -8,19 +8,19 @@ typedef struct vector vector;
 
 vector* vec(size_t type_size);
 
-vector* vec_with_capacity(size_t type_size, unsigned int capacity);
+vector* vec_with_capacity(size_t type_size, size_t capacity);
 
-vector* vec_from_array(size_t type_size, unsigned int length, const void *array);
+vector* vec_from_array(size_t type_size, size_t length, const void *array);
 
 void vec_free(vector *v);
 
-unsigned int vec_length(const vector *v);
+size_t vec_length(const vector *v);
 
-unsigned int vec_capacity(const vector *v);
+size_t vec_capacity(const vector *v);
 
 bool vec_empty(const vector *v);
 
-void* vec_at(const vector *v, unsigned int index);
+void* vec_at(const vector *v, size_t index);
 
 void* vec_begin(const vector *v);
 
@@ -30,10 +30,10 @@ void* vec_push(vector *v, const void *value);
 
 void* vec_pop(vector *v);
 
-void* vec_insert(vector *v, unsigned int index, const void *value);
+void* vec_insert(vector *v, size_t index, const void *value);
 
-void* vec_erase(vector *v, unsigned int index);
+void* vec_erase(vector *v, size_t index);
 
-void* vec_resize(vector *v, unsigned int capacity);
+void* vec_resize(vector *v, size_t capacity);
 
 #endif // INCLUDE_VECTOR_H_
